requirements-cloudrun.txt
flask==3.0.0
flask-cors==4.0.0
google-cloud-firestore==2.21.0
firebase-admin==7.1.0
requests==2.32.5
gunicorn==21.2.0